requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#	production that isn't in development.
-r base.txt
gunicorn==17.5

dj-database-url==0.2.1
django-pylibmc-sasl==0.2.4